Fixed minor bug in for duelist, App.js formatting
The duelist's deck size object was a mess of two values that I wasn't
using properly, so I simplified to a single value and applied that
change everywhere that it's used. The deck/hand/graveyard seems to
be consistent now. I also placed the body of this app in App.js into
main element tags. Not sure if that's how proper formatting is done
in React but whatever.