Separated duelist portrait selection and data
Created a new duelist reducer so that there are two profiles that work
independantly of each other. They both use the same reducer function,
however. Created a switch to swap between the two duelist profiles/slots.
When true, the button state makes it so that all radio selections
are applied to Duelist A (left), while the false value makes changes to
Duelist B (right).